refactor: move ImagePreloadManager to core-utils
