+++ core-commonui/lemmyui/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/lemmyui/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.di
+
+import com.github.diegoberaldin.raccoonforlemmy.core.commonui.lemmyui.FabNestedScrollConnection
+
+expect fun getFabNestedScrollConnection(): FabNestedScrollConnection
