--- core-commonui/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/image/ImagePreloadManager.kt
@@ -1,6 +1,0 @@
-package com.github.diegoberaldin.raccoonforlemmy.core.commonui.image
-
-interface ImagePreloadManager {
-    fun preload(url: String)
-}
-
