--- core-commonui/src/iosMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/DefaultImagePreloadManager.kt
@@ -1,10 +1,0 @@
-package com.github.diegoberaldin.raccoonforlemmy.core.commonui
-
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.image.ImagePreloadManager
-
-class DefaultImagePreloadManager() : ImagePreloadManager {
-
-    override fun preload(url: String) {
-        // no-op
-    }
-}
