+++ core-utils/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/utils/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.core.utils.di
+
+import org.koin.core.module.Module
+
+expect val imagePreloadModule: Module
