+++ core-utils/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/utils/imagepreload/ImagePreloadManager.kt
@@ -1,0 +1,6 @@
+package com.github.diegoberaldin.raccoonforlemmy.core.utils.imagepreload
+
+interface ImagePreloadManager {
+    fun preload(url: String)
+}
+
