+++ core-utils/src/iosMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/utils/imagepreload/DefaultImagePreloadManager.kt
@@ -1,0 +1,8 @@
+package com.github.diegoberaldin.raccoonforlemmy.core.utils.imagepreload
+
+class DefaultImagePreloadManager() : ImagePreloadManager {
+
+    override fun preload(url: String) {
+        // no-op
+    }
+}
